ProjectMachineLearning/app-engine-pipeline/requirements.txt
Flask==3.0.0
google-api-core==2.10.0
google-api-python-client==2.60.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-cloud-language==2.5.2
google-cloud-speech==2.15.1
google-cloud-texttospeech==2.12.1
google-cloud-automl==2.8.1
google-cloud-vision==3.1.2
google-cloud-aiplatform==1.22.1
googlemaps==4.6.0
gunicorn==20.1.0
pandas==1.4.2
requests==2.27.1
markupsafe==2.1.1
jinja2==3.1.2

